pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.3.2",
    "bcrypt>=4.3.0",
    "cryptography>=45.0.7",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.49.1",
]
